+++ lib/settings_ui.dart
@@ -4,4 +4,3 @@ export 'package:settings_ui/src/settings_section.dart';
 export 'package:settings_ui/src/settings_tile.dart';
 export 'package:settings_ui/src/settings_list.dart';
 export 'package:settings_ui/src/custom_section.dart';
-