+++ lib/src/defines.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/widgets.dart';
+
+const defaultTitlePadding = EdgeInsets.only(
+  left: 15.0,
+  right: 15.0,
+  bottom: 6.0,
+);