+++ lib/src/settings_list.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:settings_ui/src/abstract_section.dart';
 import 'package:settings_ui/src/colors.dart';
-import 'package:settings_ui/src/settings_section.dart';
 
 class SettingsList extends StatelessWidget {
   final bool shrinkWrap;
@@ -23,30 +22,32 @@ class SettingsList extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      color: Theme.of(context).brightness == Brightness.light
-          ? backgroundColor ?? lightBackgroundColor ?? backgroundGray
-          : backgroundColor ?? darkBackgroundColor ?? Colors.black,
-      child: ListView.builder(
-        physics: physics,
-        shrinkWrap: shrinkWrap,
-        itemCount: sections.length,
-        itemBuilder: (context, index) {
-          AbstractSection current = sections[index];
-          AbstractSection futureOne;
-          try {
-            futureOne = sections[index + 1];
-          } catch (e) {}
+    return Material(
+      child: Ink(
+        color: Theme.of(context).brightness == Brightness.light
+            ? backgroundColor ?? lightBackgroundColor ?? backgroundGray
+            : backgroundColor ?? darkBackgroundColor ?? Colors.black,
+        child: ListView.builder(
+          physics: physics,
+          shrinkWrap: shrinkWrap,
+          itemCount: sections.length,
+          itemBuilder: (context, index) {
+            AbstractSection current = sections[index];
+            AbstractSection futureOne;
+            if (index + 1 != sections.length) {
+              futureOne = sections[index + 1];
+            }
 
-          // Add divider if title is null
-          if (futureOne != null && futureOne.title != null) {
-            current.showBottomDivider = false;
-            return current;
-          } else {
-            current.showBottomDivider = true;
-            return current;
-          }
-        },
+            // Add divider if title is null
+            if (futureOne != null && futureOne.title != null) {
+              current.showBottomDivider = false;
+              return current;
+            } else {
+              current.showBottomDivider = true;
+              return current;
+            }
+          },
+        ),
       ),
     );
   }